Globally refactored and added localpackagelist support to (hopefully) fix client download bug.

+++ src/Wizard.Application/DirectoryManipulator/DirectoryManipulatorWindows.cs
@@ -1,45 +1,40 @@
 ﻿using System.Runtime.Versioning;
-using System.Text.RegularExpressions;
 
 namespace Wizard.Application.DirectoryManipulator;
 
 [SupportedOSPlatform("Windows")]
 public sealed class DirectoryManipulatorWindows(string directoryInstallation) : IDirectoryManipulator
 {
-    private readonly string _directoryInstallation = directoryInstallation;
     private const string PathToVersion = "Bin/data.dat";
 
     public async Task<string?> GetWizard101Version()
     {
-        var fileExists = File.Exists(Path.Combine(_directoryInstallation, PathToVersion));
+        var fileExists = File.Exists(Path.Combine(directoryInstallation, PathToVersion));
 
         if (!fileExists) return null;
-        
-        var textContent = await File.ReadAllTextAsync(Path.Combine(_directoryInstallation, PathToVersion));
+
+        var textContent = await File.ReadAllTextAsync(Path.Combine(directoryInstallation, PathToVersion));
         return textContent.Replace("\n", string.Empty);
     }
 
     public void DestroyGameData()
     {
-        var gameDataFolder = Path.Combine(_directoryInstallation, "Data", "GameData");
-        if (Directory.Exists(gameDataFolder))
-        {
-            Directory.Delete(gameDataFolder, true);
-        }
-        
-        var objectCacheFolder = Path.Combine(_directoryInstallation, "Data", "ObjectCache");
-        if (Directory.Exists(objectCacheFolder))
-        {
-            Directory.Delete(objectCacheFolder, true);
-        }
+        var gameDataFolder = Path.Combine(directoryInstallation, "Data", "GameData");
+        if (Directory.Exists(gameDataFolder)) Directory.Delete(gameDataFolder, true);
+
+        var objectCacheFolder = Path.Combine(directoryInstallation, "Data", "ObjectCache");
+        if (Directory.Exists(objectCacheFolder)) Directory.Delete(objectCacheFolder, true);
     }
 
-    public void DestroyLocalPackagesList()
+    public async Task OverrideLocalPackagesListAsync(List<string> packages)
     {
-        var localPackageListLocation = Path.Combine(_directoryInstallation, "LocalPackagesList.txt");
-        if (File.Exists(localPackageListLocation))
-        {
-            File.Delete(localPackageListLocation);
-        }
-}
+        var localPackageListLocation = Path.Combine(directoryInstallation, "LocalPackagesList.txt");
+        if (File.Exists(localPackageListLocation)) File.Delete(localPackageListLocation);
+
+        var packageStrings = packages.Select(se => se[..se.LastIndexOf('.')]).ToList();
+        packageStrings.RemoveAll(se => se.Contains("ZoneData/") || se.Contains("Shaders/"));
+        packageStrings.RemoveAll(se => se.Contains(".utd"));
+
+        await File.WriteAllLinesAsync(localPackageListLocation, packageStrings);
+    }
 }
